Fix bug with readme to named file generation
The default behavior for GitHub markdown when a link does not specify a
file directly is to display the README for that directory (if it is
available). This fixes the condition when a source link referenced a
readme this way, but the destination is now a named file.

Example:
  Document: /foo-list.md
  /foo-list.md contains a link to sig-bar as: /sig-bar
  sig-bar file path: /sig-bar/README.md
  Destination file path: /groups/sig-bar.md

Links referencing '/sig-bar' will be updated to reference the
destination path ('/groups/sig-bar.md') correctly.