Allow leading spaces for reference links
Some of the sites that get pulled in use a mix of inline and reference
links. Reference links are expected to have the format of the link
reference name in brackets, starting at the beginning of the line,
followed by a colon and then the linked target.

Some of the external sites have been found to have some variation on
this format though. While markdown references show the format as above,
it does work when these link references have leading spacing before the
opening bracket. This has caused [broken slack
links](https://github.com/kubernetes/community/issues/6934), as one
example.

This change updates the regex for identifying these reference links in
the get-content script to allow for leading spaces.

Signed-off-by: Sean McGinnis <stmcg@amazon.com>