Spotlight for CNCF DHHWG (#530)
* Create cncf-dhhwg-spotlight.md

Adding spotlight for cncf dhhwg

* Adding images for the spotlight

* Update cncf-dhhwg-spotlight.md

Fixing a few minor things

* Update cncf-dhhwg-spotlight.md

Marking a line as bold

* Update cncf-dhhwg-spotlight.md

Updating the slug

* Update cncf-dhhwg-spotlight.md

Correcting header for Wrapping up

* Update cncf-dhhwg-spotlight.md

Correcting headings to follow the Docs style guide and remove capitalisation of headings.

* Update content/en/blog/2024/cncf-dhhwg/cncf-dhhwg-spotlight.md

Incorporating suggestions

Co-authored-by: Tim Bannister <tim@scalefactory.com>

* Update content/en/blog/2024/cncf-dhhwg/cncf-dhhwg-spotlight.md

Incorporating suggestions

Co-authored-by: Tim Bannister <tim@scalefactory.com>

* Update cncf-dhhwg-spotlight.md

* Update cncf-dhhwg-spotlight.md

Updated the introduction to better connect the blog’s theme with Kubernetes, based on feedback from Tim. The focus is on linking the importance of inclusivity and contributions from underrepresented communities to the growth of Kubernetes.

* Update cncf-dhhwg-spotlight.md

* Update cncf-dhhwg-spotlight.md

Changing the publishing date to Sep 30t

---------

Co-authored-by: Tim Bannister <tim@scalefactory.com>